+++ bug-tracker.js
@@ -5,6 +5,11 @@
 (function() {
 'use strict';
 
+if (typeof window !== 'undefined' && window.APP_TEST_MODE) {
+    console.log('[BUG-TRACKER] Skipped in test mode');
+    return;
+}
+
 var STORAGE_KEY = 'tts_bug_reports';
 
 /**