+++ playwright.config.js
@@ -22,9 +22,7 @@ module.exports = defineConfig({
         '--disable-dev-shm-usage',
         '--no-sandbox',
         '--disable-setuid-sandbox',
-        '--disable-gpu',
-        '--single-process',
-        '--no-zygote'
+        '--disable-gpu'
       ]
     }
   },